requirements.txt
fastapi==0.96.0
python-dotenv==1.0.0
uvicorn==0.22.0
bcrypt==3.2.2
pytest==7.3.2
slack-sdk==3.21.3
beautifulsoup4==4.12.2
SQLAlchemy==1.4.49
pydantic==1.10.8
python-jose==3.3.0
pandas==2.0.1
Pillow==10.0.0
mplfinance==0.12.10b0